+++ loess/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Loess Changelog
 
+## next-minor
+
+TODO: Date
+
+### Features
+
+- Added: [`Input::peek(&self, )`](https://docs.rs/loess/0.1/loess/struct.Input.html#method.peek)
+- Added: [`impl PeekFrom for RArrow`](https://docs.rs/loess/0.1/loess/rust_grammar/struct.RArrow.html#impl-PeekFrom-for-RArrow) (`->`)
+
+### Revisions
+
+Various small documentation additions.
+
 ## 0.1
 
 2025-04-16